UPDATED Generate Markdown function to implement file filtering.
