+++ tsconfig.json
@@ -3,7 +3,8 @@
 		"module": "Node16",
 		"target": "ES2022",
 		"lib": [
-			"ES2022"
+			"ES2022",
+			"dom"
 		],
 		"sourceMap": true,
 		"rootDir": "src",
